gitignore 설정, 코드 단순화

+++ src/main/java/com/tradeagent/evaluation/ScoreCalculator.java
@@ -1,8 +1,6 @@
 package com.tradeagent.evaluation;
 
-import java.math.BigDecimal;
-
 public interface ScoreCalculator<T> {
 
-    BigDecimal calculate(T input);
+    double calculate(T input);
 }
